+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/model/RtcStats.kt
@@ -1,0 +1,3 @@
+package com.instrumentisto.medea_flutter_webrtc.model
+
+data class RtcStats()
