+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/model/RtcStatsReport.kt
@@ -1,0 +1,55 @@
+package com.instrumentisto.medea_flutter_webrtc.model
+
+/**
+ * Representation of an [org.webrtc.RTCStats].
+ *
+ * @property type `type` of this [RtcStats].
+ * @property id `id` of this [RtcStats].
+ * @property timestampUs `timestampUs` of this [RtcStats].
+ * @property kind map of stats of this [RtcStats].
+ */
+data class RtcStats(
+    val type: String,
+    val id: String,
+    val timestampUs: Long,
+    val kind: Map<String, Any>
+) {
+  companion object {
+    /**
+     * Converts the provided [org.webrtc.RTCStats] into an [RtcStats].
+     *
+     * @return [RtcStats] created based on the provided [org.webrtc.RTCStats].
+     */
+    fun fromWebRtc(stats: org.webrtc.RTCStats): RtcStats {
+      var members = stats.members
+      return RtcStats(stats.type, stats.id, stats.timestampUs.toLong(), members)
+    }
+  }
+  
+  /** Converts this [RtcStatsReport] into a [Map] which can be returned to the Flutter side. */
+  fun asFlutterResult(): Map<String, Any> {
+    return mapOf("id" to id, "timestampUs" to timestampUs, "kind" to kind, "type" to type)
+  }
+}
+
+/**
+ * Representation of an [org.webrtc.RTCStatsReport].
+ *
+ * @property stats `stats` of this [RtcStatsReport].
+ */
+data class RtcStatsReport(val stats: Map<String, RtcStats>) {
+  companion object {
+    /**
+     * Converts the provided [org.webrtc.RTCStatsReport] into an [RtcStatsReport].
+     *
+     * @return [RtcStatsReport] created based on the provided [org.webrtc.RTCStatsReport].
+     */
+    fun fromWebRtc(report: org.webrtc.RTCStatsReport): RtcStatsReport {
+      return RtcStatsReport(report.statsMap.mapValues { RtcStats.fromWebRtc(it.value) })
+    }
+  }
+  /** Converts this [RtcStatsReport] into a [Map] which can be returned to the Flutter side. */
+  fun asFlutterResult(): List<Map<String, Any>> {
+    return stats.map { it.value.asFlutterResult() }
+  }
+}
